MEAM: split and reorganize files by topic; unify comment style
